[#22] refactor : MVC 클래스 별 패키지 생성
MVC 클래스 별 패키지 생성
