--- BE/src/main/java/com/codesquad/baseball09/GameRepository.java
@@ -1,5 +1,0 @@
-package com.codesquad.baseball09;
-
-public interface GameRepository {
-
-}
