+++ BE/src/main/java/com/codesquad/baseball09/repository/GameRepository.java
@@ -1,0 +1,5 @@
+package com.codesquad.baseball09.repository;
+
+public interface GameRepository {
+
+}
